refactored dataExtractor, moved downloadImageData() to generic.py, works
